+++ packages/uni-app/uni_modules/vite-plugin/js_sdk/index.d.ts
@@ -2,5 +2,5 @@ export { DateFormatOptions, checkSourceExists, copySourceToTarget, deepMerge, en
 export { V as Validator } from './shared/vite-plugin.CiHfwMiN.js';
 export { BasePlugin, BasePluginOptions, OptionsNormalizer, PluginFactory, PluginWithInstance, createPluginFactory } from './factory/index.js';
 export { L as Logger, P as PluginLogger } from './shared/vite-plugin.CLr0ttuO.js';
-export { CopyFileOptions, GenerateRouterOptions, GenerateVersionOptions, Icon, InjectIcoOptions, NameStrategy, OutputFormat, OutputType, RouteConfig, RouteMeta, UniAppPageConfig, UniAppPagesJson, UniAppTabBarConfig, VersionFormat, copyFile, generateRouter, generateVersion, injectIco } from './plugins/index.js';
+export { BuildPhase, BuildProgressOptions, CopyFileOptions, GenerateRouterOptions, GenerateVersionOptions, Icon, InjectIcoOptions, NameStrategy, OutputFormat, OutputType, ProgressFormat, ProgressTheme, RouteConfig, RouteMeta, UniAppPageConfig, UniAppPagesJson, UniAppTabBarConfig, VersionFormat, buildProgress, copyFile, generateRouter, generateVersion, injectIco } from './plugins/index.js';
 import 'vite';